fixed bug in exprPow. Abort on dsyev interface errors.

+++ Couenne/src/convex/CouenneCutGenerator.cpp
@@ -81,7 +81,7 @@ int CouenneCutGenerator::addSegment (OsiCuts &cs, int wi, int xi,
   if (fabs (x2-x1) < COUENNE_EPS) {
     if (fabs (y2-y1) > MAX_SLOPE * COUENNE_EPS)
       printf ("warning, discontinuity of %e over an interval of %e\n", y2-y1, x2-x1);
-    else return createCut (cs, y2, 0, wi, 1.);
+    else return createCut (cs, y2, (int) 0, wi, 1.);
   }
 
   //CouNumber oppslope = (y1-y2) / (x2-x1);
